Secure the services with TLS (for authentication) and ACL (for authorization)

+++ internal/log/store_test.go
@@ -54,7 +54,7 @@ func testReadAt(t *testing.T, s *store) {
 		require.Equal(t, lenWidth, n)
 		off += int64(n)
 
-		size := enc.Uint16(b)
+		size := enc.Uint64(b)
 		b = make([]byte, size)
 		n, err = s.ReadAt(b, off)
 		require.NoError(t, err)
